Fix Bugs and errors in report3

+++ src/Application/Admin/Users/Controllers/UsersController.php
@@ -7,6 +7,7 @@ use Domain\User\Actions\UserGetAllAction;
 use Domain\User\Actions\UserUpdateAction;
 use Domain\User\DataTransferObjects\UserData;
 use Domain\User\Models\User;
+use Domain\Webinar\Models\Webinar;
 use Illuminate\Support\Facades\Log;
 use Spatie\Permission\Models\Role;
 
@@ -46,6 +47,9 @@ class UsersController extends \Core\Http\Controllers\Controller
 
     public function destroy(User $user)
     {
+        if ($user->hasRole('Master') && Webinar::where('master_id', $user->id)->first())
+            return back()->with('failed' , 'کاربر '. $user->name .'در تعدادی از وبینار ها نقش استاد را دارد ');
+
         $user->delete();
         return back()->with('success', 'کاربر '. $user->name .'  حذف شد.');
     }
